# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VeKay Solar CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /technicians:
    get:
      operationId: listTechnicians
      tags: [technicians]
      summary: List all technicians
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Technician"

    post:
      operationId: createTechnician
      tags: [technicians]
      summary: Create a technician (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/{id}:
    get:
      operationId: getTechnician
      tags: [technicians]
      summary: Get technician details with assigned complaints
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Technician with complaints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianDetail"

    patch:
      operationId: updateTechnician
      tags: [technicians]
      summary: Update technician
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"

    delete:
      operationId: deleteTechnician
      tags: [technicians]
      summary: Delete a technician (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /technicians/{id}/map:
    get:
      operationId: getTechnicianMap
      tags: [technicians]
      summary: Get technician assigned complaints for map view
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Map data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MapPin"

  # ── Dashboard / Summary ──────────────────────────────────────────────────
...
components:
  schemas:
...
    Complaint:
      type: object
      required: [id, ticketId, customerName, customerPhone, customerEmail, placeName, district, pincode, address, complaintType, description, status, urgency, createdAt, updatedAt]
      properties:
        id:
          type: integer
        ticketId:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        ksebConsumerNumber:
          type: ["string", "null"]
        placeName:
          type: string
        district:
          type: string
        pincode:
          type: string
        address:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        locationSource:
          type: ["string", "null"]
        complaintType:
          type: string
        description:
          type: string
        mediaUrls:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, in_progress, going, reached, on_hold, resolved, closed]
        urgency:
          type: string
          enum: [low, medium, high, critical]
        urgencyColor:
          type: ["string", "null"]
        technicianId:
          type: ["integer", "null"]
        technicianName:
          type: ["string", "null"]
        pendingTechnicianIds:
          type: array
          items:
            type: integer
        adminNotes:
          type: ["string", "null"]
        completionNotes:
          type: ["string", "null"]
        scheduledDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Technician:
      type: object
      required: [id, username, name, phone, area, isActive, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        phone:
          type: string
        area:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    TechnicianDetail:
      type: object
      required: [id, username, name, phone, area, isActive, createdAt, assignedComplaints]
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        phone:
          type: string
        area:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        assignedComplaints:
          type: array
          items:
            $ref: "#/components/schemas/Complaint"
    TechnicianInput:
      type: object
      required: [username, password, name, phone, area]
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
        phone:
          type: string
        area:
          type: string
    TechnicianUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        area:
          type: string
        isActive:
          type: boolean
        password:
          type: ["string", "null"]
    MapPin:
      type: object
      required: [complaintId, ticketId, customerName, placeName, address, lat, lng, urgency, urgencyColor, status, complaintType, createdAt]
      properties:
        complaintId:
          type: integer
        ticketId:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        placeName:
          type: string
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        urgency:
          type: string
        urgencyColor:
          type: string
        status:
          type: string
        complaintType:
          type: string
        description:
          type: string
        scheduledDate:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        isPending:
          type: boolean
        createdAt:
          type: string
...